+++ Samples~/ActorTemplate/Scripts/Data/ActorData.cs
@@ -7,17 +7,28 @@ using DSC.Event;
 
 namespace DSC.Template.Actor.Default
 {
-    public class ActorData : BaseActorData
+    public class ActorData : BaseActorData, IActorData_Input
     {
+        #region Variable
+
+        #region Variable - Property
+
+        public BaseActorInput baseInput { get; protected set; }
+
+        public EventCallback<(InputButtonType, GetInputType), BaseActorController> inputButtonCallback { get; protected set; }
+
+
+        #endregion
+
         public DSC_ActorController m_hController;
         public DSC_ActorStatus m_hStatus;
 
-        public BaseActorInput m_hInput;
-
         public ActorStateFlag m_eStateFlag;
-        public EventCallback<(InputButtonType, GetInputType), ActorData, List<IActorBehaviourData>> m_hInputButtonCallback;
 
         public float m_fDeltaTime;
+
+
+        #endregion
 
         public override void Init(Transform hActor)
         {
@@ -29,6 +40,12 @@ namespace DSC.Template.Actor.Default
             m_hController = hActor.GetComponent<DSC_ActorController>();
 
             m_hStatus = m_hActor.GetComponent<DSC_ActorStatus>();
+        }
+
+        public void InitInput(BaseActorInput hBaseInput)
+        {
+            baseInput = hBaseInput;
+            inputButtonCallback = new EventCallback<(InputButtonType, GetInputType), BaseActorController>();
         }
     }
 }