+++ Scripts/Behaviours/2D/Player/DSC_ActorBehaviour_Jump2D_Player.cs
@@ -55,19 +55,19 @@ namespace DSC.Actor.Behaviour2D
 
         public override void OnStartBehaviour(BaseActorController hBaseController)
         {
-            if (hBaseController.TryGetActorData(out BaseActorData2D_Player hActorData) 
+            if (hBaseController.TryGetIActorData(out IActorData_Input hDataInput) 
                 && hBaseController.TryGetBehaviourData(out JumpCacheData hOutData))
             {
-                hActorData.m_hInputButtonCallback.Add((m_eButton, GetInputType.Down), hOutData.m_actJump);
+                hDataInput.inputButtonCallback.Add((m_eButton, GetInputType.Down), hOutData.m_actJump);
             }
         }
 
         public override void OnStopBehaviour(BaseActorController hBaseController)
         {
-            if (hBaseController.TryGetActorData(out BaseActorData2D_Player hActorData)
+            if (hBaseController.TryGetIActorData(out IActorData_Input hDataInput)
                 && hBaseController.TryGetBehaviourData(out JumpCacheData hOutData))
             {
-                hActorData.m_hInputButtonCallback.Remove((m_eButton, GetInputType.Down), hOutData.m_actJump);
+                hDataInput.inputButtonCallback.Remove((m_eButton, GetInputType.Down), hOutData.m_actJump);
             }
         }