+++ Scripts/Behaviours/2D/Player/DSC_ActorBehaviour_WallJump2D_Player.cs
@@ -68,21 +68,21 @@ namespace DSC.Actor.Behaviour2D
 
         public override void OnStartBehaviour(BaseActorController hBaseController)
         {
-            if (hBaseController.TryGetActorData(out BaseActorData2D_Player hActorData)
+            if (hBaseController.TryGetIActorData(out IActorData_Input hDataInput)
                 && hBaseController.TryGetBehaviourData(out WallJumpCacheData hOutData))
             {
-                hActorData.m_hInputButtonCallback.Add((m_eButton, GetInputType.Down), hOutData.m_actJump);
-                hActorData.m_hInputButtonCallback.Add((m_eButton, GetInputType.Up), hOutData.m_actJumpUp);
+                hDataInput.inputButtonCallback.Add((m_eButton, GetInputType.Down), hOutData.m_actJump);
+                hDataInput.inputButtonCallback.Add((m_eButton, GetInputType.Up), hOutData.m_actJumpUp);
             }
         }
 
         public override void OnStopBehaviour(BaseActorController hBaseController)
         {
-            if (hBaseController.TryGetActorData(out BaseActorData2D_Player hActorData)
+            if (hBaseController.TryGetIActorData(out IActorData_Input hDataInput)
                 && hBaseController.TryGetBehaviourData(out WallJumpCacheData hOutData))
             {
-                hActorData.m_hInputButtonCallback.Remove((m_eButton, GetInputType.Down), hOutData.m_actJump);
-                hActorData.m_hInputButtonCallback.Remove((m_eButton, GetInputType.Up), hOutData.m_actJumpUp);
+                hDataInput.inputButtonCallback.Remove((m_eButton, GetInputType.Down), hOutData.m_actJump);
+                hDataInput.inputButtonCallback.Remove((m_eButton, GetInputType.Up), hOutData.m_actJumpUp);
             }
         }
 
@@ -130,7 +130,7 @@ namespace DSC.Actor.Behaviour2D
         void OnJump(BaseActorController hBaseController)
         {
             if (hBaseController.isTimeStop || !PassCondition(hBaseController)
-                || !hBaseController.TryGetActorData(out BaseActorData2D_Player hActorData)
+                || !hBaseController.TryGetActorData(out BaseActorData2D hActorData)
                 || !hBaseController.TryGetBehaviourData(out WallJumpCacheData hOutData)
                 || hOutData.m_bWallJumping)
                 return;