+++ Scripts/Cores/BaseActorData.cs
@@ -9,16 +9,10 @@ namespace DSC.Actor
     public class BaseActorData : ScriptableObject
     {
         public Transform m_hActor;
-        public BaseActorInput m_hInput;
-
-        public EventCallback<(InputButtonType, GetInputType), BaseActorController> m_hInputButtonCallback;
-
 
         public virtual void Init(Transform hActor)
         {
             m_hActor = hActor;
-
-            m_hInputButtonCallback = new EventCallback<(InputButtonType, GetInputType), BaseActorController>();
         }
     }
 
@@ -34,15 +28,6 @@ namespace DSC.Actor
                 return;
 
             m_hPhysic = hActor.GetComponent<BaseActorPhysic2D>();
-        }
-    }
-
-    public class BaseActorData2D_Player : BaseActorData2D
-    {
-        
-        public override void Init(Transform hActor)
-        {
-            base.Init(hActor);            
         }
     }
 }