+++ Scripts/Interface/IActorData/IActorData_Input.cs
@@ -1,0 +1,14 @@
+﻿using DSC.Core;
+using DSC.Event;
+
+namespace DSC.Actor
+{
+    public interface IActorData_Input : IActorData
+    {
+        BaseActorInput baseInput { get; }
+
+        EventCallback<(InputButtonType, GetInputType), BaseActorController> inputButtonCallback { get; }
+
+        void InitInput(BaseActorInput hBaseInput);
+    }
+}